requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cftime==1.6.4.post1
Django==5.0
django-filter==25.1
django-postgres-copy==2.8.0
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
drf-yasg==1.21.10
inflection==0.5.1
netCDF4==1.7.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
xarray==2025.7.1
